move GHOST mete keys to plugin
