+++ tests/plugins/ghost/test_meta_keys.py
@@ -4,7 +4,7 @@ Metadata keys from GHOST that only depend on station dimension
 THESE SHOULD BE ALWAYS THE SAME ACCORDING TO INFORMATION INFORMATION
 PROVIDED BY BSC
 """
-from pyaerocom.io.ghost_meta_keys import GHOST_META_KEYS
+from pyaerocom.plugins.ghost.meta_keys import GHOST_META_KEYS
 
 desired_keys = [
     "EDGAR_v4.3.2_annual_average_BC_emissions",
